Split tests for grundsteuer input.
